requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
pymilvus
sentence-transformers
groq
langgraph
langchain
langchain-groq
pymupdf
python-docx
pydantic
python-dotenv
werkzeug